+++ src/editor/components/menu/editor-mode-title.jsx
@@ -70,7 +70,4 @@ export function mapDispatchToProps(dispatch) {
     }
   };
 }
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(TitleUnconnected);
+export default connect(mapStateToProps, mapDispatchToProps)(TitleUnconnected);